+++ components/hero/homepage-hero.js
@@ -1,5 +1,4 @@
 import Image from "next/image";
-import Typewriter from "typewriter-effect";
 import growthImage from "../../public/images/growth-2.png";
 
 export default function HomepageHero() {
@@ -17,15 +16,9 @@ export default function HomepageHero() {
 				<h1 className="z-10 max-w-4xl font-display text-5xl font-medium tracking-tight text-left text-slate-900 sm:text-7xl">
 					Lack the drive to level up as a developer? You’re in the right place!
 				</h1>
-				<div className="p-4 mx-auto lg:mx-0 mt-16 max-w-2xl text-2xl tracking-tight text-left">
-					<Typewriter
-						options={{
-							strings: ["Make learning, levelling up fun and interactive!"],
-							autoStart: true,
-							loop: true,
-						}}
-					/>
-				</div>
+				<p className="p-4 mx-auto lg:mx-0 mt-16 max-w-2xl text-2xl tracking-tight text-left">
+					Make learning, levelling up fun and interactive!
+				</p>
 			</div>
 		</section>
 	);